requirements.txt
rasa
openai
python-dotenv
black